Fix crash when playing long sounds (like DSBOSSIT)

+++ src/i_sdlsound.c
@@ -143,7 +143,7 @@ static void ExpandSoundData(byte *data,
 
         // number of samples in the converted sound
 
-        expanded_length = (length * mixer_freq) / samplerate;
+        expanded_length = ((uint64_t) length * mixer_freq) / samplerate;
         expand_ratio = (length << 8) / expanded_length;
 
         for (i=0; i<expanded_length; ++i)
